requirements.txt
solders
solana
base58
requests
azure-identity
azure-keyvault-secrets
python-dotenv